Bad implementation of limits (fixed cardinality))
